requirements.txt
streamlit
google-generativeai
ffmpeg-python